requirements.txt
pydantic>=2.0
requests
python-dotenv
tavily-python
langchain
langchain-core